% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aca}
\alias{aca}
\title{Albo degli Arbitri della Camera Arbitrale}
\format{
\subsection{\code{aca}}{

A data frame with 234 rows and 7 columns:
\describe{
\item{cognome, nome}{Arbitrator's surname and name}
\item{data_nascita}{Arbitrator's birthday}
\item{descrizione}{Arbitrator's occupation description }
\item{nr_iscrizione_albo}{Arbitrator's register id}
\item{data_iscrizione_albo, data_cessazione_albo}{ Professional registration date and termiantion}
...
}
}
}
\source{
\url{https://dati.anticorruzione.it/#/aca}
}
\usage{
aca
}
\description{
Data regarding official arbitrators' register people
}
\keyword{datasets}
