% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_index}
\alias{check_index}
\title{Check the "index" in xml tree}
\usage{
check_index(raw)
}
\description{
This function checks wether the xml supplied
is and "index" or not. Index xmls groups tenders
data by year.
}
