% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dati-comunicazione-opec.R
\name{parse_xml_dati}
\alias{parse_xml_dati}
\title{parse_xml_dati}
\usage{
parse_xml_dati(xml_url, annoDiComunicazione)
}
\arguments{
\item{xml_url}{url that points to xml file}

\item{annoDiComunicazione}{PARAM_DESCRIPTIO}
}
\value{
xml tibble dataframe
}
\description{
It parses xml dati
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  parse_xml_dati("http://www.halleyweb.com/ucpentri/zf/index.php/dataset/appalti-2020.xml")
}
}
}
