% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dati-comunicazione-opec.R
\name{parse_xml_index}
\alias{parse_xml_index}
\title{parse_xml_index}
\usage{
parse_xml_index(href, annoDiComunicazione)
}
\arguments{
\item{href}{href corresponding to the index annoDiComunicazione}

\item{annoDiComunicazione}{the annoDiComunicazione to which the CA has communicaterd to ANAC the xml}
}
\value{
tibble
}
\description{
it transforms and cleans index type xml file into tabular i.e. tibble dataframe class
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  parse_xml_index("http://contrattipubblici.sipalinformatica.it/l190c32/JOSWG705/indiceDataset.xml")
}
}
}
