% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{possibly2}
\alias{possibly2}
\title{possibly2}
\usage{
possibly2(.f, otherwise = NULL)
}
\arguments{
\item{.f}{A function, formula, or vector (not necessarily atomic).}

\item{otherwise}{Default function to use when an error occurs., Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
traditional \code{purrr::possibly()} with dynamic \code{otherwise} method
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
