% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcpt}
\alias{rcpt}
\title{Elenco Responsabili Prevenzione Corruzione e Trasparenza}
\format{
\subsection{\code{rcpt}}{

A data frame with 6,339 rows and 6 columns:
\describe{
\item{nome, cognome}{name and suname}
\item{data_nomina}{nominee date}
\item{link_atto_nomina}{deed of appointment's link}
\item{denominazione_persona_giurdica}{legal entity name}
\item{codice_fiscale_persona_giurdica}{legal entity id}
...
}
}
}
\source{
\url{https://dati.anticorruzione.it/#/rpct}
}
\usage{
rcpt
}
\description{
Data regarding people responsible for corruption prevention and transparency
}
\keyword{datasets}
