% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{registro_comunicazioni}
\alias{registro_comunicazioni}
\title{Registro comunicazioni Art.1 comma 32 L.190/2012}
\format{
\subsection{\code{rcpt}}{

A data frame with 6,339 rows and 6 columns:
\describe{
\item{identificativoPEC}{certified e-mail address (PEC ion ita) address }
\item{codiceFiscale}{id/vat for Certified Authority (CA)}
\item{ragioneSociale}{CA extented name}
\item{url}{xml url pointing to tender data for that year for that CA }
\item{dataUltimoTentativoAccessoUrl, esitoUltimoTentativoAccessoUrl}{latest communication date between CA and ANAC and outcome}
\item{annoDiRiferimento}{the year ANAC has received communication }
...
}
}
}
\source{
\url{https://dati.anticorruzione.it/#/l190}
}
\usage{
registro_comunicazioni
}
\description{
data List of communications received and processed
pursuant to art. 1, paragraph 32, of Law 190/2012;
}
\keyword{datasets}
