% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{serv_fatt}
\alias{serv_fatt}
\title{servizi-fatturazione-elettronica}
\format{
\subsection{\code{serv_fatt}}{

A data frame with 53,473 rows and 12 columns:
\describe{
\item{cod_amm,cod_ou,cod_uni_ou}{administrative, ou, and unique ou id}
\item{des_ou}{ou description)}
\item{regione,provincia, comune }{region,province and municipality}
\item{indirizzo, cap}{address and cap, i.e. postal code }
\item{cf,}{tax id for CA}
\item{dt_verifica_cf, data_avvio_sfe}{verification and starting service date }
...
}
}
}
\source{
\url{https://indicepa.gov.it/ipa-dati/dataset/servizi-fatturazione-elettronica}
}
\usage{
serv_fatt
}
\description{
The dataset contains information on the electronic invoicing services present in IPA
structured with the methods prior to March 31, 2021 (TXT with tabulator separator).
}
\keyword{datasets}
