% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_cf}
\alias{validate_cf}
\alias{validate_sceltaContraente}
\title{validate_cf}
\usage{
validate_cf(cf)

validate_sceltaContraente(sc)
}
\arguments{
\item{cf}{codice fiscale aplphanumeric string (valid for both PA and citizens) case insensitive}

\item{sc}{codice fiscale aplphanumeric string (valid for both PA and citizens) case insensitive}
}
\value{
bool

bool
}
\description{
Check if \code{cf} ie "codice fiscale" is valid

CCheck if \code{sc} ie "sceltaContraente" is valid.
}
\examples{
\dontrun{
if (interactive()) {
  validate_cf("slvncl96b06a564a")
  validate_cf("03179300920")
}
}
\dontrun{
if (interactive()) {
  validate_sceltaContraente("29-PROCEDURA RISTRETTA SEMPLIFICATA")
  validate_sceltaContraente("29-PROCEDURE RISTRETTA SEMPLIFICATU")
}
}
}
